+++ imageparts.go
@@ -66,12 +66,12 @@ func (w *wholeImage) Src(i int) (x0, y0, x1, y1 int) {
 	return 0, 0, w.width, w.height
 }
 
-func u(x float32, width int) float32 {
-	return x / float32(internal.NextPowerOf2Int(width))
+func u(x float64, width int) float64 {
+	return x / float64(internal.NextPowerOf2Int(width))
 }
 
-func v(y float32, height int) float32 {
-	return y / float32(internal.NextPowerOf2Int(height))
+func v(y float64, height int) float64 {
+	return y / float64(internal.NextPowerOf2Int(height))
 }
 
 type textureQuads struct {
@@ -84,13 +84,13 @@ func (t *textureQuads) Len() int {
 	return t.parts.Len()
 }
 
-func (t *textureQuads) Vertex(i int) (x0, y0, x1, y1 float32) {
+func (t *textureQuads) Vertex(i int) (x0, y0, x1, y1 float64) {
 	ix0, iy0, ix1, iy1 := t.parts.Dst(i)
-	return float32(ix0), float32(iy0), float32(ix1), float32(iy1)
+	return float64(ix0), float64(iy0), float64(ix1), float64(iy1)
 }
 
-func (t *textureQuads) Texture(i int) (u0, v0, u1, v1 float32) {
+func (t *textureQuads) Texture(i int) (u0, v0, u1, v1 float64) {
 	x0, y0, x1, y1 := t.parts.Src(i)
 	w, h := t.width, t.height
-	return u(float32(x0), w), v(float32(y0), h), u(float32(x1), w), v(float32(y1), h)
+	return u(float64(x0), w), v(float64(y0), h), u(float64(x1), w), v(float64(y1), h)
 }
